pip-requirements.txt
absl-py==1.3.0
brotlipy==0.7.0
cachetools==5.2.0
certifi
cffi
charset-normalizer==2.1.1
cox==0.1.post3
cryptography==38.0.3
cycler==0.11.0
dill==0.3.6
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.7.1
GPUtil==1.4.0
grpcio==1.50.0
idna==3.4
importlib-metadata==5.0.0
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
#mkl-fft==1.3.1
#mkl-random
#mkl-service==2.4.0
numexpr==2.8.4
numpy==1.21.6
nvidia-dali-cuda110==1.16.0
oauthlib==3.2.2
packaging==21.3
pandas==1.3.5
Pillow==9.3.0
pip-review==1.2.0
protobuf==3.19.6
psutil==5.9.4
py3nvml==0.2.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests
requests-oauthlib==1.3.1
robustness==1.2.1.post2
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.1
six
smmap==5.0.0
tables==3.7.0
tensorboard
#==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard-wrapper==0.1.3
tensorboardX==2.5.1
threadpoolctl==3.1.0
#torch==1.12.1+cu113
torchaudio==0.12.1
torchsummary==1.5.1
#torchvision==0.13.1+cu113
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
xmltodict==0.13.0
zipp==3.10.0
